同步点赞数据，增加用户是否点赞判断

+++ fresh-system/src/main/java/com/creamakers/fresh/system/domain/dto/FreshNews.java
@@ -48,7 +48,7 @@ public class FreshNews {
 
     // 点赞数量
     @TableField(value = "liked")
-    private Integer liked;
+    private Long liked;
 
     // 评论数量
     @TableField(value = "comments")
